backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-extensions==0.7.1
idna==3.4
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
